+++ packages/kit/src/runtime/client/start.js
@@ -7,15 +7,7 @@ import { set_legacy } from '../legacy.js';
 /**
  * @param {{
  *   env: Record<string, string>;
- *   hydrate: {
- *     status: number;
- *     error: Error | (import('../server/page/types').SerializedHttpError);
- *     node_ids: number[];
- *     params: Record<string, string>;
- *     routeId: string | null;
- *     data: Array<import('types').ServerDataNode | null>;
- *     errors: Record<string, any> | null;
- *   };
+ *   hydrate: Parameters<import('./types').Client['_hydrate']>[0];
  *   paths: {
  *     assets: string;
  *     base: string;