+++ packages/kit/src/runtime/server/endpoint.js
@@ -1,4 +1,5 @@
-import { Redirect } from '../control.js';
+import { json } from '../../exports/index.js';
+import { Redirect, ValidationError } from '../control.js';
 import { check_method_names, method_not_allowed } from './utils.js';
 
 /**
@@ -56,6 +57,8 @@ export async function render_endpoint(event, mod, state) {
 				status: error.status,
 				headers: { location: error.location }
 			});
+		} else if (error instanceof ValidationError) {
+			return json(error.data, { status: error.status });
 		}
 
 		throw error;