+++ packages/kit/src/runtime/server/page/respond_with_error.js
@@ -25,7 +25,8 @@ export async function respond_with_error({ event, options, state, status, error,
 		event,
 		options,
 		state,
-		route: GENERIC_ERROR
+		route: GENERIC_ERROR,
+		resolve_opts
 	});
 
 	try {
@@ -79,8 +80,7 @@ export async function respond_with_error({ event, options, state, status, error,
 			fetched,
 			cookies,
 			event,
-			resolve_opts,
-			validation_errors: undefined
+			resolve_opts
 		});
 	} catch (err) {
 		const error = coalesce_to_error(err);