+++ packages/kit/test/apps/basics/src/routes/shadowed/post-success-redirect/+page.server.js
@@ -1,5 +1,8 @@
-export function POST() {
-	return {
-		location: '/shadowed/post-success-redirect/redirected'
-	};
-}
+import { redirect } from '@sveltejs/kit';
+
+/** @type {import('./$types').Actions} */
+export const actions = {
+	default: () => {
+		throw redirect(303, '/shadowed/post-success-redirect/redirected');
+	}
+};