+++ packages/kit/test/apps/basics/test/server.test.js
@@ -277,15 +277,17 @@ test.describe('Routing', () => {
 });
 
 test.describe('Shadowed pages', () => {
-	test('Action can return undefined', async ({ request }) => {
+	test('Action can return undefined', async ({ baseURL, request }) => {
 		const response = await request.post('/shadowed/simple/post', {
+			form: {},
 			headers: {
-				accept: 'application/json'
+				accept: 'application/json',
+				origin: new URL(baseURL).origin
 			}
 		});
 
-		expect(response.status()).toBe(204);
-		expect(await response.text()).toEqual('');
+		expect(response.status()).toBe(200);
+		expect(await response.json()).toEqual({ type: 'success', status: 204 });
 	});
 });