+++ packages/kit/test/apps/dev-only/test/test.js
@@ -36,3 +36,18 @@ test.describe('$env', () => {
 		);
 	});
 });
+
+test.describe('server-only modules', () => {
+	test('server-only module is not statically importable from the client', async ({ request }) => {
+		const resp = await request.get('/server-only-modules/static-import');
+		expect(await resp.text()).toMatch(
+			/.*Error: Cannot import \$lib\/test.server.js into client-side code:.*/gs
+		);
+	});
+	test('server-only module is not dynamically importable from the client', async ({ request }) => {
+		const resp = await request.get('/server-only-modules/dynamic-import');
+		expect(await resp.text()).toMatch(
+			/.*Error: Cannot import \$lib\/test.server.js into client-side code:.*/gs
+		);
+	});
+});